Removed the use of mfuse in public include files, using std containers instead to reduce the load of dependencies Removed the use of *Vector* class from public include files, using vec*X*_t instead, because not everyone has the same math library
Using Eigen as a vector math library

Assets:
- Moved SkeletorManager from utility to assets

Network:
- using std::chrono for network time, much better and more precise

Network/CGame:
- Moved prediction code to another class
- Removed instance of TraceManager from cgame (instance it manually instead)
- Removed prediction settings, using a boolean parameter for prediction enabled

Network/Client:
- Created a class for encoding/decoding packets
- Created InputModule class for anything related to user inputs
- Moved channel-related code from ServerConnection to a ServerChannel class
- Removed ping code from ServerSnapshot
- Fixed time adjustment
- Fixed user input being late

Network/Files:
- Moved FillGameDirectory code out of the class, made it an helper function instead so that it's more generic

Network/Parsing:
- Added input parsing class

Network/Remote:
- Prevent processing packets from another remote in netchan, only process packets from our remote connection
- New classes for simulating UDP packet latency, loss, duplication and out-of-order

Network/Types:
- Refactored user commands
- Added new classes for buttons and weapon commands.

Utility:
- Added dynamic enumeration
- Added WarpArray

Utility/Misc:
- Renamed EndianHelpers to EndianCoordHelpers

Utility/Misc/MSG:
- Moved MsgTypesHelper to MsgCoordHelper in a new file

Tests:
- Split code between network and assets

+++ include/MOHPC/Network/Parsing/GameState.h
@@ -47,7 +47,7 @@ namespace Network
 			gameStateResults_t();
 
 		public:
-			uint64_t serverDeltaTime;
+			uint32_t serverDeltaTime;
 			float serverDeltaTimeSeconds;
 			uint32_t serverId;
 		};
